+++ library/src/main/java/com/evernote/android/job/v21/JobProxy21.java
@@ -200,6 +200,8 @@ public class JobProxy21 implements JobProxy {
                 return JobInfo.NETWORK_TYPE_UNMETERED;
             case NOT_ROAMING:
                 return JobInfo.NETWORK_TYPE_UNMETERED; // use unmetered here, is overwritten in v24
+            case METERED:
+                return JobInfo.NETWORK_TYPE_ANY; // use any here as fallback
             default:
                 throw new IllegalStateException("not implemented");
         }