+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Remove the persisted parameter, which didn't work reliable, all jobs are persisted anyway
 * Remove `startWakefulService` from the `Job` class, `WakefulBroadcastReceiver` is now deprecated with `JobIntentService` as the better option
 * Add feature to make jobs transient and to add a `Bundle`, see `setTransientExtras(bundle)`
+* Add new `METERED` network type
 
 ## 1.1.12